+++ pkg/traceql/engine.go
@@ -269,7 +269,9 @@ func (e *Engine) asTraceSearchMetadata(spanset *Spanset) *tempopb.TraceSearchMet
 				attribute.Intrinsic == IntrinsicDuration ||
 				attribute.Intrinsic == IntrinsicTraceDuration ||
 				attribute.Intrinsic == IntrinsicTraceRootService ||
-				attribute.Intrinsic == IntrinsicTraceRootSpan {
+				attribute.Intrinsic == IntrinsicTraceRootSpan ||
+				attribute.Intrinsic == IntrinsicTraceID ||
+				attribute.Intrinsic == IntrinsicSpanID {
 				continue
 			}